+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ master
 1) [BUGFIX] Fix wrong namespace.
 2) [TASK] Use more strict comparison.
 3) [TASK] Add ddev.
+4) [TASK][BREAKING] Increase dependency to `sourcebroker/deployer-instance`.
 
 7.1.0
 ~~~~~~